requirements.txt
Flask
Flask-Cors
requests
python-dotenv
python-docx
PyMuPDF
Pillow
pytesseract 